ISDB-317: Implementation of c100 transformation
