+++ src/integrationTest/resources/request/bulk-scan-c100-transform-input.json
@@ -263,6 +263,90 @@
     {
       "name": "natureofOrder",
       "value": "yes"
+    },
+    {
+      "name": "NoMIAM_DVE_relevantPoliceCaution",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_relevantCriminalProceeding",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_relevantConviction",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_courtOrder",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_protectionNotice",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_protectiveInjunction",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_NoCrossUndertakingGiven",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_copyOfFactFinding",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_expertEvidenceReport",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_healthProfessionalReport",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_ReferralHealthProfessionalReport",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_memberOf_MultiAgencyRiskAssessmentConferrance_letter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_domesticViolenceAdvisor",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_independentSexualViolenceAdvisor_Letter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_officerEmployed_localAuthority_letter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_domesticViolenceSupportCharity_letter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_domesticViolenceSupportCharity_refuge_letter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_publicAuthority_confirmationLetter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_secretaryOfState_letter",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_evidenceFinancialMatters",
+      "value": "true"
+    },
+    {
+      "name": "NoMIAM_DVE_arrestedForSimilarOffence",
+      "value": "true"
     }
   ],
   "scanned_documents": [
