+++ src/main/java/uk/gov/hmcts/reform/bulkscan/services/BulkScanC100Service.java
@@ -5,8 +5,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.bulkscan.config.BulkScanFormValidationConfigManager;
 import uk.gov.hmcts.reform.bulkscan.config.BulkScanTransformConfigManager;
-import uk.gov.hmcts.reform.bulkscan.enums.ChildLiveWithEnum;
-import uk.gov.hmcts.reform.bulkscan.enums.PermissionRequiredEnum;
 import uk.gov.hmcts.reform.bulkscan.helper.BulkScanTransformHelper;
 import uk.gov.hmcts.reform.bulkscan.helper.BulkScanValidationHelper;
 import uk.gov.hmcts.reform.bulkscan.model.BulkScanTransformationRequest;
@@ -23,8 +21,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static org.apache.commons.lang3.BooleanUtils.TRUE;
-import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.APPLICATION_PERMISSION_REQUIRED;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.BULK_SCAN_CASE_REFERENCE;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CASE_TYPE_ID;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILDREN_OF_SAME_PARENT;
@@ -38,7 +34,6 @@ import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.CHILD_LOC
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.EVENT_ID;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.MANDATORY_ERROR_MESSAGE;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.NO;
-import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.PERMISSION_REQUIRED;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.XOR_CONDITIONAL_FIELDS_MESSAGE;
 import static uk.gov.hmcts.reform.bulkscan.constants.BulkScanConstants.YES;
 
@@ -139,11 +134,7 @@ public class BulkScanC100Service implements BulkScanService {
             .transformToCaseData(new HashMap<>(transformConfigManager.getTransformationConfig(FormType.C100)
                     .getCaseDataFields()), inputFieldsMap);
 
-        populatedMap.put(CHILD_LIVE_WITH_KEY, getChildLiveWith(inputFieldsMap));
-        populatedMap.put(APPLICATION_PERMISSION_REQUIRED, getApplicationPermissionRequired(inputFieldsMap));
-
-        populatedMap.put(SCAN_DOCUMENTS, transformScanDocuments(bulkScanTransformationRequest));
-
+        bulkScanC100ConditionalTransformerService.transform(populatedMap, inputFieldsMap, bulkScanTransformationRequest);
         Map<String, String> caseTypeAndEventId =
             transformConfigManager.getTransformationConfig(FormType.C100).getCaseFields();
 
@@ -153,31 +144,5 @@ public class BulkScanC100Service implements BulkScanService {
                 .caseTypeId(caseTypeAndEventId.get(CASE_TYPE_ID))
                 .eventId(caseTypeAndEventId.get(EVENT_ID))
                 .caseData(populatedMap).build()).build();
-    }
-
-    private String getChildLiveWith(Map<String, String> inputFieldsMap) {
-        if (TRUE.equalsIgnoreCase(inputFieldsMap.get(CHILD_LIVING_WITH_APPLICANT))) {
-            return ChildLiveWithEnum.APPLICANT.getName();
-        }
-        if (TRUE.equals(inputFieldsMap.get(CHILD_LIVING_WITH_RESPONDENT))) {
-            return ChildLiveWithEnum.RESPONDENT.getName();
-        }
-        if (TRUE.equals(inputFieldsMap.get(CHILD_LIVING_WITH_OTHERS))) {
-            return ChildLiveWithEnum.OTHERPEOPLE.getName();
-        }
-        return StringUtils.EMPTY;
-    }
-
-    private String getApplicationPermissionRequired(Map<String, String> inputFieldsMap) {
-        if (YES.equalsIgnoreCase(inputFieldsMap.get(PERMISSION_REQUIRED))) {
-            return PermissionRequiredEnum.yes.getDisplayedValue();
-        }
-        if ("No, permission Not required".equals(inputFieldsMap.get(PERMISSION_REQUIRED))) {
-            return PermissionRequiredEnum.noNotRequired.getDisplayedValue();
-        }
-        if ("No, permission Now sought".equals(inputFieldsMap.get(PERMISSION_REQUIRED))) {
-            return PermissionRequiredEnum.noNowSought.getDisplayedValue();
-        }
-        return StringUtils.EMPTY;
     }
 }
